requirements/dev.txt
black==26.1.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
djlint==1.36.4
dnspython==2.8.0
EditorConfig==0.17.1
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.1
idna==3.11
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.13.0
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.1.15
six==1.17.0
SQLAlchemy==2.0.46
tqdm==4.67.1
typing_extensions==4.15.0
Werkzeug==3.1.5
WTForms==3.2.1
